Code rebasing
Rebasing of the code of the component, including major changes
Changes in the naming of the topics used for the communication of the component
Better information propagation during the creation of a Broker subscriber processing function
Better handling of upper and lower metric (attribute) bounds
Acquisition of broker credential information from the user
Improvement of number handling within metrics (allow a metric to have integer or real values)
Improvement of the logic related to the initiation of a new SLO violation detection engine
Transferral of generic topic subscriptions (application-independent logic) to the code of the Director Subcomponent
Improvements to the CustomDataPublisher
Better handling of paths for the properties files necessary for the operation of the component
Improvement on the handling of infinity
Implementation of stopping functionality for AMQP connectors through ExtendedConnector
Changes by specifying explicitly the names of topics
Introduction of GUI subscriber utility and improvements on GUI publisher
Improvements in Debug data generated
Changes in REST calls to use application name
Improvements in the unregistration of consumers and stopping of an ExtendedConnector
Improvements in the initialization of the component
Addition of capability to export executable jar
Fixed a case when the active connector could not be stopped because it was not non-null.
Removed unecessary code, and jar-generating code.
Allow setting custom port, and set the correct host to contact the activemq broker

Change-Id: I94a6fdb4612de192c24511445f1236cdce94b402